football_league_manager/requirements.txt
beautifulsoup4==4.12.3
Flask==3.0.2
SQLAlchemy==2.0.28
pymysql==1.1.0
mysql==0.0.3
mysql-connector-python==8.3.0
Flask-SQLAlchemy==3.1.1
Flask-Migrat==4.0.7